+++ src/main/java/com/abz/task_manager/services/impl/TaskServiceImpl.java
@@ -7,8 +7,8 @@ import com.abz.task_manager.domain.entities.TaskStatus;
 import com.abz.task_manager.repositories.TaskListRepositories;
 import com.abz.task_manager.repositories.TaskRepository;
 import com.abz.task_manager.services.TaskService;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
@@ -47,7 +47,7 @@ public class TaskServiceImpl implements TaskService {
         TaskPriority taskPriority = Optional.ofNullable(task.getPriority()).orElse(TaskPriority.MEDIUM);
         TaskStatus taskStatus = TaskStatus.OPEN;
         TaskList taskList = taskListRepositories.findById(taskListId)
-                .orElseThrow(() -> new IllegalArgumentException("Task list does not exist"));
+                .orElseThrow(() -> new EntityNotFoundException("Task list does not exist"));
 
 
         return taskRepository.save(new Task(null, task.getTitle(), task.getDescription(), task.getDueDate(), taskStatus, taskPriority, now, now, taskList));
@@ -62,28 +62,21 @@ public class TaskServiceImpl implements TaskService {
 
     @Override
     public Task updateTask(UUID taskListId, UUID taskId, Task task) {
-        if (task.getId() == null) {
-            throw new IllegalArgumentException("Task id is required");
-        }
-
-        if(task.getId() != taskId){
+        if (task.getId() != null && !task.getId().equals(taskId)) {
             throw new IllegalArgumentException("changing task id not allowed");
-
-
         }
-        if(task.getTitle() == null || task.getTitle().isBlank()) {
+        if (task.getTitle() == null || task.getTitle().isBlank()) {
             throw new IllegalArgumentException("Task title is required");
-
-
         }
-        if(task.getPriority() == null) {
+        if (task.getPriority() == null) {
             throw new IllegalArgumentException("Task priority is required");
         }
-        if(task.getStatus() == null) {
+        if (task.getStatus() == null) {
             throw new IllegalArgumentException("Task status is required");
         }
 
-        Task existingTask = getTask(taskListId, taskId).orElseThrow(() -> new IllegalArgumentException("Task does not exist"));
+        Task existingTask = getTask(taskListId, taskId)
+                .orElseThrow(() -> new EntityNotFoundException("Task does not exist"));
         LocalDateTime now = LocalDateTime.now();
         existingTask.setTitle(task.getTitle());
         existingTask.setDescription(task.getDescription());
@@ -97,6 +90,9 @@ public class TaskServiceImpl implements TaskService {
     @Transactional
     @Override
     public void deleteTask(UUID taskListId, UUID id) {
+        if (taskRepository.findByTaskListIdAndId(taskListId, id).isEmpty()) {
+            throw new EntityNotFoundException("Task does not exist");
+        }
         taskRepository.deleteByTaskListIdAndId(taskListId, id);
     }
 }
